fix(qlegal): Ask was broken on prod — 'operator is not unique: date + unknown'
Regression I introduced when wiring the Ask business-rule params: the upcoming-
obligations query became 'current_date + $1'. Postgres cannot infer a bare
parameter's type there — date + integer and date + interval are both candidates —
so it rejects the whole query and every Ask request failed. Cast explicitly to
$1::int. Verified against Postgres directly, not just by syntax check.

Also darkens the two grey tokens at source in BOTH Q-Legal and Contra
(--dim #54504A→#3D3934, --dim2 #7A7266→#5C564D) — one change, every use.

+++ server/qlegal.js
@@ -561,7 +561,9 @@ export function mountQLegal(app, upload) {
       )).rows;
       const soon = (await q(
         `select d.filename, o.kind, o.what, o.due_date from ql_obligation o join ql_document d on d.id=o.document_id
-          where o.status in ('proposed','confirmed') and o.due_date between current_date and current_date + $1 order by o.due_date limit 15`, [AP.obligations_horizon_days]
+          where o.status in ('proposed','confirmed')
+            and o.due_date between current_date and current_date + ($1::int)
+          order by o.due_date limit 15`, [AP.obligations_horizon_days]
       )).rows;
       rungs.push("C2/REGISTERS (structured, whole estate)");
